Fix linting error

+++ src/Helpers/Tools.php
@@ -103,7 +103,7 @@ class Tools
     public static function humanFileSize($bytes, $decimals = 2): string
     {
         $size = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-        $factor = floor((strlen((string) $bytes) - 1) / 3);
+        $factor = (int) floor((strlen((string) $bytes) - 1) / 3);
 
         return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)).@$size[$factor];
     }
